Fixed conflicts:
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_EInvalidInput_decl.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_EInvalidInput_def.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_EPhysicalProblem_decl.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_EPhysicalProblem_def.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_EReadError_decl.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_EReadError_def.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_ESetupError_decl.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_ESetupError_def.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_MSSMNoFV_onshell_soft_parameters_decl.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_MSSMNoFV_onshell_soft_parameters_def.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_MSSMNoFV_onshell_susy_parameters_decl.hpp
	Backends/include/gambit/Backends/backend_types/gm2calc_1_3_0/wrapper_MSSMNoFV_onshell_susy_parameters_def.hpp
	Backends/include/gambit/Backends/default_bossed_versions.hpp
	Backends/include/gambit/Backends/frontends/SuperIso_3_6.hpp
	DarkBit/include/gambit/DarkBit/ProcessCatalog.hpp
	Elements/include/gambit/Elements/printable_types.hpp
	FlavBit/include/gambit/FlavBit/FlavBit_rollcall.hpp
	FlavBit/src/FlavBit.cpp
	Printers/include/gambit/Printers/printers/hdf5types.hpp
	Utils/include/gambit/Utils/util_functions.hpp
	cmake/backends.cmake
	cmake/scanners.cmake
	cmake/utilities.cmake
	config/capabilities.dat
	yaml_files/CMSSM.yaml

+++ Utils/include/gambit/Utils/yaml_node_utility.hpp
@@ -50,7 +50,7 @@ namespace Gambit
             os << "Provided value " << d << " as option in the yaml file does not represent an integer.";
             utils_error().raise(LOCAL_INFO, os.str());
           }
-          if (std::numeric_limits<TYPE>::max() < d or std::numeric_limits<TYPE>::min() > d)
+          if (static_cast<double>(std::numeric_limits<TYPE>::max()) < d or static_cast<double>(std::numeric_limits<TYPE>::min()) > d)
           {
             std::ostringstream os;
             os << "Provided value " << d << " as option in the yaml file does not fit into the implemented integer type.";
